Add digital formats to holdings: ignore list
Although format-specific index partials prevent display of the "On order" message on results pages, addition of this message to the catalogue page (PR87) means that it displays for (digital) formats with no holding records. The "ignore" lists for each environment prevent this evaluation: "Video" and "E-Thesis" (forthcoming) have been added to each.