Place subjects on new line and terminate with '.'
UAT feedback on CER‌-0009 highlighted the value of hyphenating subject terms and presenting each on a separate line. A further request to terminate each subject field with a full stop is also accommodated in the display_field_search_link helper.
From a search of the code in the blah repository, only the subject display field uses this method, but the change would obviously affect any future use for other data.