Hyphenate subject terms at the subfield
UAT feedback on CER‌-0009 highlighted the value of hyphenating subject terms. The custom indexing method used for subject_t does not have the option of specifying a seperator other than the default space.  However, getAllSubfields does. If this method is used when indexing subject_t, with the hyphenation string set as a parameter, the desired result is achieved.